Changed config to fix file formating, replaced js for jsx files where needed
