+++ neighbors/.eslintrc
@@ -27,6 +27,7 @@
     "react/jsx-props-no-spreading": "off",
     "react-hooks/exhaustive-deps": "warn",
     "react/display-name": "off",
+    "react/react-in-jsx-scope": "off",
 
     "import/no-default-export": "error",
     "import/order": [
