--- neighbors/src/App.js
@@ -1,13 +1,0 @@
-/* eslint-disable react/jsx-no-undef */
-/* eslint-disable prettier/prettier */
-import React from 'react';
-
-import './App.css';
-import { Registration } from './components/registration';
-
-export const App = () => {
-  return(
-  <>
-    <Registration></Registration>
-  </>)
-};
