+++ neighbors/src/App.jsx
@@ -1,0 +1,10 @@
+import './App.css';
+import { Registration } from './components/registration';
+
+export const App = () => {
+  return (
+    <>
+      <Registration></Registration>
+    </>
+  );
+};
