+++ neighbors/src/api/firebase.config.js
@@ -1,10 +1,9 @@
-/* eslint-disable prettier/prettier */
 export const firebaseConfig = {
-apiKey: 'AIzaSyBY6p3bjo3T_xw7rItYjg8krzjH57R8h0o',
-    authDomain: 'neighbors-17a98.firebaseapp.com',
-    projectId: 'neighbors-17a98',
-storageBucket: 'neighbors-17a98.appspot.com',
-    messagingSenderId: "958727918389",
-    appId: "1:958727918389:web:b81a7e562f5e29c6e9f27c",
-    measurementId: "G-B8MRC2G602"
+  apiKey: 'AIzaSyBY6p3bjo3T_xw7rItYjg8krzjH57R8h0o',
+  authDomain: 'neighbors-17a98.firebaseapp.com',
+  projectId: 'neighbors-17a98',
+  storageBucket: 'neighbors-17a98.appspot.com',
+  messagingSenderId: '958727918389',
+  appId: '1:958727918389:web:b81a7e562f5e29c6e9f27c',
+  measurementId: 'G-B8MRC2G602',
 };
